Evaluar Menu e ingresar comentarios

+++ src/main/java/com/proyecto/transferObject/EvaluacionTO.java
@@ -1,41 +1,49 @@
 package com.proyecto.transferObject;
 
 public class EvaluacionTO {
-	
+
 	private int id;
-	private int valoracion;
+	private float valoracion;
 	private String comentario;
 	private int idMenu;
 	private int idUsuario;
-	
-	
+
 	public int getId() {
 		return id;
 	}
+
 	public void setId(int id) {
 		this.id = id;
 	}
-	public int getValoracion() {
+
+	public float getValoracion() {
 		return valoracion;
 	}
-	public void setValoracion(int valoracion) {
+
+	public void setValoracion(float valoracion) {
 		this.valoracion = valoracion;
 	}
+
 	public String getComentario() {
 		return comentario;
 	}
+
 	public void setComentario(String comentario) {
 		this.comentario = comentario;
 	}
+
 	public int getIdMenu() {
 		return idMenu;
 	}
+
 	public void setIdMenu(int idMenu) {
 		this.idMenu = idMenu;
 	}
+
 	public int getIdUsuario() {
 		return idUsuario;
 	}
+
 	public void setIdUsuario(int idUsuario) {
 		this.idUsuario = idUsuario;
 	}
